# docs/api-specs/openapi/quizmentor-api-v1.yaml
openapi: 3.0.3
info:
  title: QuizMentor API
  description: |
    Complete REST API specification for QuizMentor - A gamified quiz learning platform.
    This API provides endpoints for authentication, quiz management, gamification, 
    multiplayer features, and analytics.
  version: 1.0.0
  contact:
    name: QuizMentor Team
    email: api@quizmentor.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3002/api
    description: Local development server
  - url: https://api.quizmentor.com
    description: Production server
  - url: https://staging-api.quizmentor.com
    description: Staging server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /quiz/session:
    post:
      tags: [Quiz]
      summary: Create quiz session
      operationId: createQuizSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuizRequest'
      responses:
        '201':
          description: Quiz session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizSession'
...
  /leaderboard:
    get:
      tags: [Gamification]
      summary: Get leaderboard
      operationId: getLeaderboard
      security: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [global, friends, category, weekly, monthly]
        - name: category_id
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Leaderboard data
          content:
            application/json:
              schema:
                type: object
                properties:
                  leaderboard:
                    type: array
                    items:
                      $ref: '#/components/schemas/LeaderboardEntry'
                  user_rank:
                    $ref: '#/components/schemas/LeaderboardEntry'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    CreateQuizRequest:
      type: object
      required: [category_id, question_count]
      properties:
        category_id:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced, expert, adaptive]
        question_count:
          type: integer
          minimum: 1
          maximum: 50
        mode:
          type: string
          enum: [practice, timed, multiplayer, challenge, survival]
        time_per_question:
          type: integer
          minimum: 10
          maximum: 300
    QuizSession:
      type: object
      required: [id, status, questions, config]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, active, paused, completed, abandoned]
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
        current_question:
          type: integer
        config:
          type: object
          properties:
            total_questions:
              type: integer
            time_per_question:
              type: integer
            scoring_method:
              type: string
              enum: [standard, speed_bonus, accuracy_bonus, combo]
            difficulty_mode:
              type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    Question:
      type: object
      required: [id, text, options, difficulty, category]
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
        code_snippet:
          type: string
        image_url:
          type: string
          format: uri
        options:
          type: array
          minItems: 2
          maxItems: 6
          items:
            type: object
            required: [id, text]
            properties:
              id:
                type: string
              text:
                type: string
              image_url:
                type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced, expert]
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        time_limit:
          type: integer
        points:
          type: integer
...
    LeaderboardEntry:
      type: object
      required: [rank, user, score]
      properties:
        rank:
          type: integer
        user:
          type: object
          required: [id, display_name]
          properties:
            id:
              type: string
            display_name:
              type: string
            avatar_url:
              type: string
        score:
          type: integer
        xp:
          type: integer
        level:
          type: integer
        streak:
          type: integer
        badges:
          type: array
          items:
            type: string
        trend:
          type: string
          enum: [up, down, stable, new]
...